+++ BetterSplitscreen/Layout/Layout.cs
@@ -1,0 +1,7 @@
+﻿namespace BetterSplitscreen.Layout
+{
+    internal class Layout
+    {
+
+    }
+}
