+++ BetterSplitscreen/ModConfig.cs
@@ -1,0 +1,6 @@
+﻿namespace BetterSplitscreen
+{
+    internal class ModConfig
+    {
+    }
+}
